+++ build.config.ts
@@ -1,13 +1,14 @@
-// eslint-disable-next-line import/no-extraneous-dependencies
 import { defineBuildConfig } from "unbuild";
 
 export default defineBuildConfig({
   entries: ["./src/index"],
   clean: true,
   outDir: "lib",
+  declaration: false,
   rollup: {
+    inlineDependencies: true,
     esbuild: {
-      minify: true
+      minify: false
     }
   }
 });